implemented utils file with functions to calculate sample mean and std dev

+++ calculator/src/main/java/calculator/CatSample.java
@@ -1,5 +1,9 @@
 package calculator;
 
+/**
+ * This class represents a categorical sample with a sample size, sample proportion, sample standard deviation
+ * @author Noah Lago ndl3389@rit.edu
+ */
 public class CatSample {
     /**
      * Represents the size of the sample instance
@@ -10,9 +14,9 @@ public class CatSample {
      * Represents the proportion of this specific sample, not of overall population
      * Common notation for sample proportion is "p̂"
      */
-    private int sp;
+    private double sp;
 
-    public CatSample(int n, int sp){
+    public CatSample(int n, double sp){
         this.n = n;
         this.sp = sp;
     }
@@ -21,7 +25,7 @@ public class CatSample {
         return n;
     }
 
-    public int getSp() {
+    public double getSp() {
         return sp;
     }
 }
